Add files via upload
I am currently in the learning of HTML and CSS, these are my small projects that I am creating and learning as I go. I welcome good and flaw feedbacks on my post so any comment is okay, I wan to know more of what I can work on to better myself in web developing. If any new self taught are looking for someone to do small practice projects I am open that as well. Thanks!